requirements.txt
python-telegram-bot[socks]>=21.7
python-dotenv>=1.0.1
lark-oapi>=1.5.3
